Add custom modules for ModelVersion log_model.
